Put API in tables, add type information
I had to dig into the source code to see what the types and defaults were for the API. This should be documented. I put the data into markdown tables, to make it easier to read. This is something I do for all of my own libraries. However if you don't like the tables, that's fine, but the data still needs to be displayed in some format in the docs.